Resolve PR suggestions.
