+++ metrics/cli/lib/cli/deployer/factory/deployer_factory.dart
@@ -2,8 +2,8 @@
 // that can be found in the LICENSE file.
 
 import 'package:cli/cli/deployer/deployer.dart';
-import 'package:cli/common/model/factory/paths_factory.dart';
-import 'package:cli/common/model/factory/services_factory.dart';
+import 'package:cli/common/model/paths/factory/paths_factory.dart';
+import 'package:cli/common/model/services/factory/services_factory.dart';
 import 'package:cli/prompter/prompter.dart';
 import 'package:cli/prompter/writer/io_prompt_writer.dart';
 import 'package:cli/util/file/file_helper.dart';
