+++ metrics/cli/lib/common/model/config/firebase_config.dart
