+++ metrics/cli/lib/common/model/config/sentry_web_config.dart
