+++ metrics/cli/lib/common/model/paths/factory/paths_factory.dart
@@ -1,7 +1,7 @@
 // Use of this source code is governed by the Apache License, Version 2.0
 // that can be found in the LICENSE file.
 
-import 'package:cli/common/model/paths.dart';
+import 'package:cli/common/model/paths/paths.dart';
 
 /// A class providing method for creating a [Paths] instance.
 class PathsFactory {
