+++ metrics/cli/test/cli/deployer/factory/deployer_factory_test.dart
@@ -3,7 +3,7 @@
 
 import 'package:cli/cli/deployer/deployer.dart';
 import 'package:cli/cli/deployer/factory/deployer_factory.dart';
-import 'package:cli/common/model/services.dart';
+import 'package:cli/common/model/services/services.dart';
 import 'package:mockito/mockito.dart';
 import 'package:test/test.dart';
