+++ metrics/cli/test/cli/doctor/doctor_test.dart
@@ -2,7 +2,7 @@
 // that can be found in the LICENSE file.
 
 import 'package:cli/cli/doctor/doctor.dart';
-import 'package:cli/common/model/services.dart';
+import 'package:cli/common/model/services/services.dart';
 import 'package:mockito/mockito.dart';
 import 'package:test/test.dart';
