+++ metrics/cli/test/cli/updater/config/factory/update_config_factory_test.dart
@@ -1,15 +1,16 @@
 // Use of this source code is governed by the Apache License, Version 2.0
 // that can be found in the LICENSE file.
 
-import 'dart:io';
-
-import 'package:cli/cli/updater/config/model/update_config.dart';
-import 'package:cli/cli/updater/config/parser/update_config_parser.dart';
 import 'package:cli/cli/updater/config/factory/update_config_factory.dart';
-import 'package:cli/util/file/file_helper.dart';
+import 'package:cli/cli/updater/config/parser/update_config_parser.dart';
+import 'package:cli/common/model/config/firebase_config.dart';
+import 'package:cli/common/model/config/sentry_config.dart';
+import 'package:cli/common/model/config/update_config.dart';
 import 'package:mockito/mockito.dart';
 import 'package:test/test.dart';
 
+import '../../../../test_utils/file_helper_mock.dart';
+import '../../../../test_utils/file_mock.dart';
 import '../../../../test_utils/matchers.dart';
 
 // ignore_for_file: avoid_redundant_argument_values
@@ -19,26 +20,26 @@ void main() {
     const configPath = 'configPath';
     const content = 'content';
 
-    final fileHelper = _FileHelperMock();
+    final fileHelper = FileHelperMock();
     final configParser = _UpdateConfigParserMock();
-    final file = _FileMock();
+    final file = FileMock();
     final configFactory = UpdateConfigFactory(
       fileHelper: fileHelper,
       configParser: configParser,
     );
+
+    PostExpectation<String> whenReadAsString({String withConfigPath}) {
+      final path = withConfigPath ?? configPath;
+
+      when(fileHelper.getFile(path)).thenReturn(file);
+      return when(file.readAsStringSync());
+    }
 
     tearDown(() {
       reset(fileHelper);
       reset(configParser);
       reset(file);
     });
-
-    PostExpectation<String> whenReadContent({String withConfigPath}) {
-      final path = withConfigPath ?? configPath;
-
-      when(fileHelper.getFile(path)).thenReturn(file);
-      return when(file.readAsStringSync());
-    }
 
     test(
       "creates an instance with the given parameters",
@@ -87,7 +88,7 @@ void main() {
     test(
       ".create() gets the configuration file using the given file helper",
       () {
-        whenReadContent().thenReturn(content);
+        whenReadAsString().thenReturn(content);
 
         configFactory.create(configPath);
 
@@ -96,9 +97,9 @@ void main() {
     );
 
     test(
-      ".create() reads the configuration file",
+      ".create() reads the content of the configuration file as string",
       () {
-        whenReadContent().thenReturn(content);
+        whenReadAsString().thenReturn(content);
 
         configFactory.create(configPath);
 
@@ -109,7 +110,7 @@ void main() {
     test(
       ".create() parses the config file content using the given config parser",
       () {
-        whenReadContent().thenReturn(content);
+        whenReadAsString().thenReturn(content);
 
         configFactory.create(configPath);
 
@@ -120,24 +121,24 @@ void main() {
     test(
       ".create() returns the config parsed by the update config parser",
       () {
-        const json = {
-          'firebase': {
-            'auth_token': 'token',
-            'project_id': 'projectId',
-            'google_sign_in_client_id': 'clientId',
-          },
-          'sentry': {
-            'auth_token': 'token',
-            'organization_slug': 'orgSlug',
-            'project_slug': 'projectSlug',
-            'project_dsn': 'projectDsn',
-            'release_name': 'releaseName'
-          },
-        };
+        final firebaseConfig = FirebaseConfig(
+          authToken: 'token',
+          projectId: 'projectId',
+          googleSignInClientId: 'google_sign_in_client_id',
+        );
+        final sentryConfig = SentryConfig(
+          authToken: 'token',
+          organizationSlug: 'organizationSlug',
+          projectSlug: 'projectSlug',
+          projectDsn: 'projectDsn',
+          releaseName: 'releaseName',
+        );
+        final expectedConfig = UpdateConfig(
+          firebaseConfig: firebaseConfig,
+          sentryConfig: sentryConfig,
+        );
 
-        final expectedConfig = UpdateConfig.fromJson(json);
-
-        whenReadContent().thenReturn(content);
+        whenReadAsString().thenReturn(content);
         when(configParser.parse(content)).thenReturn(expectedConfig);
 
         final config = configFactory.create(configPath);
@@ -148,8 +149,4 @@ void main() {
   });
 }
 
-class _FileHelperMock extends Mock implements FileHelper {}
-
 class _UpdateConfigParserMock extends Mock implements UpdateConfigParser {}
-
-class _FileMock extends Mock implements File {}
