+++ metrics/cli/test/cli/updater/config/parser/update_config_parser_test.dart
@@ -1,9 +1,13 @@
 // Use of this source code is governed by the Apache License, Version 2.0
 // that can be found in the LICENSE file.
 
-import 'package:cli/cli/updater/config/model/update_config.dart';
 import 'package:cli/cli/updater/config/parser/update_config_parser.dart';
+import 'package:cli/common/model/config/update_config.dart';
+import 'package:mockito/mockito.dart';
 import 'package:test/test.dart';
+import 'package:yaml_map/yaml_map.dart';
+
+import '../../../../test_utils/matchers.dart';
 
 void main() {
   group("UpdateConfigParser", () {
@@ -33,10 +37,30 @@ void main() {
     };
     const configParser = UpdateConfigParser();
 
+    final yamlMapParserMock = YamlMapParserMock();
+
+    tearDown(() {
+      reset(yamlMapParserMock);
+    });
+
     test(
       ".parse() throws an ArgumentError if the given config yaml is null",
       () {
         expect(() => configParser.parse(null), throwsArgumentError);
+      },
+    );
+
+    test(
+      ".parse() parses the given content using the YAML map parser",
+      () {
+        final configParser = UpdateConfigParser(
+          parser: yamlMapParserMock,
+        );
+        final content = json.toString();
+
+        configParser.parse(content);
+
+        verify(yamlMapParserMock.parse(content)).called(once);
       },
     );
 
@@ -52,3 +76,5 @@ void main() {
     );
   });
 }
+
+class YamlMapParserMock extends Mock implements YamlMapParser {}
