+++ metrics/cli/test/common/model/config/update_config_test.dart
@@ -1,0 +1,115 @@
+// Use of this source code is governed by the Apache License, Version 2.0
+// that can be found in the LICENSE file.
+
+import 'package:cli/common/model/config/firebase_config.dart';
+import 'package:cli/common/model/config/sentry_config.dart';
+import 'package:cli/common/model/config/update_config.dart';
+import 'package:test/test.dart';
+
+// ignore_for_file: avoid_redundant_argument_values
+
+void main() {
+  group("UpdateConfig", () {
+    const firebaseAuthToken = 'firebaseAuthToken';
+    const projectId = 'projectId';
+    const googleSignInClientId = 'googleSignInClientId';
+    const sentryAuthToken = 'sentryAuthToken';
+    const organizationSlug = 'organizationSlug';
+    const projectSlug = 'projectSlug';
+    const projectDsn = 'projectDsn';
+    const releaseName = 'releaseName';
+    const firebaseJson = {
+      'auth_token': firebaseAuthToken,
+      'project_id': projectId,
+      'google_sign_in_client_id': googleSignInClientId,
+    };
+    const sentryJson = {
+      'auth_token': sentryAuthToken,
+      'organization_slug': organizationSlug,
+      'project_slug': projectSlug,
+      'project_dsn': projectDsn,
+      'release_name': releaseName,
+    };
+    const json = {
+      'firebase': firebaseJson,
+      'sentry': sentryJson,
+    };
+
+    final firebaseConfig = FirebaseConfig(
+      authToken: firebaseAuthToken,
+      projectId: projectId,
+      googleSignInClientId: googleSignInClientId,
+    );
+    final sentryConfig = SentryConfig(
+      authToken: sentryAuthToken,
+      organizationSlug: organizationSlug,
+      projectSlug: projectSlug,
+      projectDsn: projectDsn,
+      releaseName: releaseName,
+    );
+
+    test(
+      "throws an ArgumentError if the given firebase config is null",
+      () {
+        expect(
+          () => UpdateConfig(
+            firebaseConfig: null,
+          ),
+          throwsArgumentError,
+        );
+      },
+    );
+
+    test(
+      "creates an instance with the given parameters",
+      () {
+        final config = UpdateConfig(
+          firebaseConfig: firebaseConfig,
+          sentryConfig: sentryConfig,
+        );
+
+        expect(config.firebaseConfig, equals(firebaseConfig));
+        expect(config.sentryConfig, equals(sentryConfig));
+      },
+    );
+
+    test(
+      ".fromJson() returns null if the given json is null",
+      () {
+        final config = UpdateConfig.fromJson(null);
+
+        expect(config, isNull);
+      },
+    );
+
+    test(
+      ".fromJson() creates an instance from the JSON-encodable map",
+      () {
+        final expected = UpdateConfig(
+          firebaseConfig: firebaseConfig,
+          sentryConfig: sentryConfig,
+        );
+
+        final config = UpdateConfig.fromJson(json);
+
+        expect(config, equals(expected));
+      },
+    );
+
+    test(
+      "equals to another UpdateConfig with the same parameters",
+      () {
+        final expected = UpdateConfig(
+          firebaseConfig: firebaseConfig,
+          sentryConfig: sentryConfig,
+        );
+        final updateConfig = UpdateConfig(
+          firebaseConfig: firebaseConfig,
+          sentryConfig: sentryConfig,
+        );
+
+        expect(updateConfig, equals(expected));
+      },
+    );
+  });
+}
