+++ metrics/cli/test/common/model/config/web_metrics_config_test.dart
@@ -1,8 +1,8 @@
 // Use of this source code is governed by the Apache License, Version 2.0
 // that can be found in the LICENSE file.
 
-import 'package:cli/common/model/sentry_web_config.dart';
-import 'package:cli/common/model/web_metrics_config.dart';
+import 'package:cli/common/model/config/sentry_web_config.dart';
+import 'package:cli/common/model/config/web_metrics_config.dart';
 import 'package:test/test.dart';
 
 void main() {
