+++ metrics/cli/test/common/model/paths/factory/paths_factory_test.dart
@@ -1,8 +1,8 @@
 // Use of this source code is governed by the Apache License, Version 2.0
 // that can be found in the LICENSE file.
 
-import 'package:cli/common/model/paths.dart';
-import 'package:cli/common/model/factory/paths_factory.dart';
+import 'package:cli/common/model/paths/paths.dart';
+import 'package:cli/common/model/paths/factory/paths_factory.dart';
 import 'package:test/test.dart';
 
 void main() {
