+++ metrics/cli/test/common/model/services/factory/services_factory_test.dart
@@ -1,7 +1,7 @@
 // Use of this source code is governed by the Apache License, Version 2.0
 // that can be found in the LICENSE file.
 
-import 'package:cli/common/model/factory/services_factory.dart';
+import 'package:cli/common/model/services/factory/services_factory.dart';
 import 'package:cli/services/firebase/adapter/firebase_cli_service_adapter.dart';
 import 'package:cli/services/flutter/adapter/flutter_cli_service_adapter.dart';
 import 'package:cli/services/gcloud/adapter/gcloud_cli_service_adapter.dart';
