+++ metrics/cli/test/common/model/services/services_test.dart
@@ -1,15 +1,15 @@
 // Use of this source code is governed by the Apache License, Version 2.0
 // that can be found in the LICENSE file.
 
-import 'package:cli/common/model/services.dart';
+import 'package:cli/common/model/services/services.dart';
 import 'package:test/test.dart';
 
-import '../../test_utils/firebase_service_mock.dart';
-import '../../test_utils/flutter_service_mock.dart';
-import '../../test_utils/gcloud_service_mock.dart';
-import '../../test_utils/git_service_mock.dart';
-import '../../test_utils/npm_service_mock.dart';
-import '../../test_utils/sentry_service_mock.dart';
+import '../../../test_utils/firebase_service_mock.dart';
+import '../../../test_utils/flutter_service_mock.dart';
+import '../../../test_utils/gcloud_service_mock.dart';
+import '../../../test_utils/git_service_mock.dart';
+import '../../../test_utils/npm_service_mock.dart';
+import '../../../test_utils/sentry_service_mock.dart';
 
 // ignore_for_file: avoid_redundant_argument_values
