+++ metrics/cli/test/test_utils/file_mock.dart
@@ -1,7 +1,8 @@
 // Use of this source code is governed by the Apache License, Version 2.0
 // that can be found in the LICENSE file.
 
-import 'package:cli/common/model/services/factory/services_factory.dart';
+import 'dart:io';
+
 import 'package:mockito/mockito.dart';
 
-class ServicesFactoryMock extends Mock implements ServicesFactory {}
+class FileMock extends Mock implements File {}
