+++ metrics/cli/test/test_utils/services_mock.dart
@@ -1,7 +1,7 @@
 // Use of this source code is governed by the Apache License, Version 2.0
 // that can be found in the LICENSE file.
 
-import 'package:cli/common/model/services.dart';
+import 'package:cli/common/model/services/services.dart';
 import 'package:mockito/mockito.dart';
 
 class ServicesMock extends Mock implements Services {}
