+++ metrics/cli/test/util/file/file_helper_test.dart
@@ -1,13 +1,12 @@
 // Use of this source code is governed by the Apache License, Version 2.0
 // that can be found in the LICENSE file.
 
-import 'dart:io';
-
 import 'package:cli/util/file/file_helper.dart';
 import 'package:mockito/mockito.dart';
 import 'package:test/test.dart';
 
 import '../../test_utils/directory_mock.dart';
+import '../../test_utils/file_mock.dart';
 import '../../test_utils/matchers.dart';
 
 void main() {
@@ -23,7 +22,7 @@ void main() {
     };
     const helper = FileHelper();
 
-    final file = _FileMock();
+    final file = FileMock();
     final directory = DirectoryMock();
 
     tearDown(() {
@@ -132,5 +131,3 @@ void main() {
     );
   });
 }
-
-class _FileMock extends Mock implements File {}
